requirements.txt
azure-core==1.0.0
azure-functions==1.0.4
azure-storage-blob==12.0.0
certifi==2019.9.11
cffi==1.13.2
chardet==3.0.4
cryptography==2.8
idna==2.8
isodate==0.6.0
msrest==0.6.10
numpy==1.17.4
oauthlib==3.1.0
onnxruntime==1.0.0
Pillow==6.2.1
pycparser==2.19
requests==2.22.0
requests-oauthlib==1.3.0
six==1.13.0
urllib3==1.25.7
